type-safety in zen_config calls
(eg: can't pass default null to a function which requires a string)

+++ admin/attributes_controller.php
@@ -1112,7 +1112,7 @@ if ($_GET['products_filter'] === 0) {
           ORDER BY LPAD(po.products_options_sort_order,11,'0'),
                    LPAD(pa.options_id,11,'0'),
                    LPAD(pa.products_options_sort_order,11,'0')";
-    $attributes_split = new splitPageResults($_GET['page'], zen_config('MAX_DISPLAY_SEARCH_RESULTS'), $attributes_query_raw, $attributes_query_numrows);
+    $attributes_split = new splitPageResults($_GET['page'], (int)zen_config('MAX_DISPLAY_SEARCH_RESULTS'), $attributes_query_raw, $attributes_query_numrows);
 ?>
         <div class="row">
           <?= zen_draw_separator('pixel_trans.gif') ?>
